reorganize imageapi module into multiple files
