+++ modules/imageapi/imageapi.go
@@ -17,29 +17,19 @@
 package imageapi
 
 import (
-	"encoding/json"
 	"fmt"
-	math "math"
 	"os"
 	"reflect"
-	"regexp"
 	"sync"
 	"time"
 
-	"github.com/davecgh/go-spew/spew"
-	transport "github.com/go-openapi/runtime/client"
-	"github.com/go-openapi/strfmt"
 	proto "github.com/gogo/protobuf/proto"
 	ptypes "github.com/gogo/protobuf/types"
 	"github.com/hpc/kraken/core"
 	cpb "github.com/hpc/kraken/core/proto"
 	ia "github.com/hpc/kraken/extensions/imageapi"
-	kjson "github.com/hpc/kraken/lib/json"
 	"github.com/hpc/kraken/lib/types"
 	"github.com/hpc/kraken/lib/util"
-	api "github.com/jlowellwofford/imageapi/client"
-	"github.com/jlowellwofford/imageapi/client/containers"
-	"github.com/jlowellwofford/imageapi/models"
 )
 
 ////////////////////////////////////
@@ -352,786 +342,4 @@ func (is *ImageAPI) Init(api types.ModuleAPIClient) {
 // Stop should perform a graceful exit
 func (is *ImageAPI) Stop() {
 	os.Exit(0)
-}
-
-////////////////////////
-// Unexported methods /
-//////////////////////
-
-////////////////////////
-// Mutation handlers //
-//////////////////////
-
-func (is *ImageAPI) mUKtoIDLE(me *core.MutationEvent) {
-	// force a discovery run and go to idle
-	is.clearAllTriggers()
-	is.mutex.Lock()
-	is.target = ia.ImageState_IDLE
-	is.mutex.Unlock()
-
-	is.discover()
-}
-
-func (is *ImageAPI) mANYtoACTIVE(me *core.MutationEvent) {
-	// Activate all Images (if they aren't already)
-	is.mutex.Lock()
-	is.target = ia.ImageState_ACTIVE
-	is.mutex.Unlock()
-	is.setRunning()
-	v, _ := is.api.QueryGetValueDsc(is.api.Self().String(), isURL)
-	isd := v.(ia.ImageSet)
-	for name, _ := range isd.Images {
-		is.activateImage(name)
-	}
-}
-
-func (is *ImageAPI) activateImage(name string) {
-	if !is.isRunning() || is.target != ia.ImageState_ACTIVE {
-		// do nothing
-		return
-	}
-	var cfgImage, dscImage ia.Image
-
-	imgURL := util.URLPush(util.URLPush(isURL, "Images"), name)
-	v, err := is.api.QueryGetValueDsc(is.api.Self().String(), imgURL)
-	if err != nil {
-		// shouldn't have been called
-		is.api.Logf(types.LLDDEBUG, "actiaveImage called on an image that has no dsc state: %s", name)
-	}
-	dscImage = v.(ia.Image)
-
-	action := dscImage.Action
-	if action == is.getAction(name) {
-		// we're already doing it...
-		is.api.Logf(types.LLDDDEBUG, "asked to perform an action we're already performing: %s on %s", action.String(), name)
-		return
-	}
-	is.api.Logf(types.LLDDEBUG, "activating %s (%s), dsc: %v\n", name, action.String(), spew.Sdump(dscImage))
-
-	if action == ia.Image_DELETE {
-		is.setAction(name, ia.Image_DELETE)
-		is.deleteImage(name, &ia.Image{})
-		is.setTrigger(name, ContainerState_DELETED, is.clearAction)
-		return
-	}
-
-	// from this point we need the cfgImage
-	v, err = is.api.QueryGetValue(is.api.Self().String(), imgURL)
-	if err != nil {
-		is.api.Logf(types.LLERROR, "asked to perform a non-delete action on a non-existent image: %s : %s", name, action.String())
-		return
-	}
-	cfgImage = v.(ia.Image)
-
-	switch action {
-	case ia.Image_CREATE:
-		is.setAction(name, ia.Image_CREATE)
-		is.createImage(name, &cfgImage)
-	case ia.Image_RELOAD:
-		is.setAction(name, ia.Image_RELOAD)
-		is.reloadImage(name, &cfgImage)
-	case ia.Image_NONE:
-		return
-	}
-	is.setTrigger(name, cfgImage.Container.State, is.tSetACTIVE)
-	is.setTrigger(name, cfgImage.Container.State, is.clearAction)
-}
-
-func (is *ImageAPI) mANYtoIDLE(me *core.MutationEvent) {
-	// Stop & delete all Images
-	is.setRunning()
-	is.mutex.Lock()
-	is.target = ia.ImageState_IDLE
-	is.mutex.Unlock()
-	v, _ := me.NodeCfg.GetValue(isURL)
-	isc := v.Interface().(ia.ImageSet)
-	v, _ = me.NodeDsc.GetValue(isURL)
-	isd := v.Interface().(ia.ImageSet)
-	for name, image := range isc.Images {
-		if _, ok := isd.Images[name]; ok {
-			// this container needs to be deleted
-			is.setAction(name, ia.Image_DELETE)
-			is.deleteImage(name, image)
-			is.setTrigger(name, ContainerState_DELETED, is.tSetIDLE)
-			is.setTrigger(name, ContainerState_DELETED, is.clearAction)
-		}
-	}
-}
-
-func (is *ImageAPI) mRecoverError(me *core.MutationEvent) bool {
-	// logic for when an error is recoverable
-	cleared := true
-	changed := false
-	is.mutex.Lock()
-	v, _ := is.api.QueryGetValueDsc(is.api.Self().String(), isURL)
-	isd := v.(ia.ImageSet)
-	for name, image := range isd.Images {
-		if image.State == ia.ImageState_ERROR {
-			image.Retries++
-			changed = true
-			if image.Retries > is.cfg.MaxRetries {
-				// too many retries
-				cleared = false
-				image.State = ia.ImageState_FATAL
-				image.LastError = ia.Image_MAX_ATTEMPTS
-				is.api.Logf(types.LLERROR, "image reached maximum retries (%d/%d): %s", image.Retries, is.cfg.MaxRetries, name)
-			} else {
-				image.State = ia.ImageState_UPDATE
-				image.Action = ia.Image_RELOAD
-				backoff := int(5 * math.Pow(2, float64(image.Retries-1))) // FIXME: this probably should be configurable
-				is.api.Logf(types.LLERROR, "image is retrying (%d/%d) in %ds: %s", image.Retries, is.cfg.MaxRetries, backoff, name)
-				time.Sleep(time.Duration(backoff) * time.Second)
-			}
-		}
-	}
-	if changed {
-		if err := is.api.QuerySetValueDsc(is.api.Self().String(), isURL, isd); err != nil {
-			is.api.Logf(types.LLERROR, "failed to set updated error value(s): %v", err)
-		}
-	}
-	is.mutex.Unlock()
-	if changed {
-		is.updateSetState()
-	}
-	return cleared
-}
-
-func (is *ImageAPI) setValues(vs map[string]interface{}) {
-	is.mutex.Lock()
-	if _, err := is.api.QuerySetValuesDsc(is.api.Self().String(), vs); err != nil {
-		is.api.Logf(types.LLERROR, "setvalues failed: %v: %v", err, vs)
-	}
-	is.mutex.Unlock()
-}
-
-/////////////////////
-// Event handlers //
-///////////////////
-
-func (is *ImageAPI) handleMutation(m types.Event) {
-	if m.Type() != types.Event_STATE_MUTATION {
-		is.api.Log(types.LLINFO, "got an unexpected event type on mutation channel")
-	}
-	me := m.Data().(*core.MutationEvent)
-
-	// mutation switch
-	switch me.Type {
-	case core.MutationEvent_MUTATE:
-		is.clearAllTriggers()
-		is.clearAllActions()
-		is.clearRunning()
-		switch me.Mutation[1] {
-		case "UKtoIDLE":
-			is.mUKtoIDLE(me)
-		case "UPDATEtoACTIVE":
-			time.Sleep(1 * time.Second) // FIXME: this is a hack to make sure more updates have accumulated
-			fallthrough
-		case "IDLEtoACTIVE":
-			is.mANYtoACTIVE(me)
-		case "ACTIVEtoIDLE", "UPDATEtoIDLE":
-			is.mANYtoIDLE(me)
-		case "ERRORtoACTIVE":
-			if is.mRecoverError(me) {
-				is.mANYtoACTIVE(me)
-			}
-		case "ERRORtoIDLE":
-			if is.mRecoverError(me) {
-				is.mANYtoIDLE(me)
-			}
-		case "SYNCtoINIT":
-			// we just directly discover the state
-			if is.pollTicker != nil {
-				is.pollTicker.Stop()
-			}
-			url := util.NodeURLJoin(is.api.Self().String(), issURL)
-			// fixme: this should really be automatically handled by the sme
-			is.dchan <- core.NewEvent(
-				types.Event_DISCOVERY,
-				url,
-				&core.DiscoveryEvent{
-					URL:     url,
-					ValueID: ia.ImageState_UNKNOWN.String(),
-				},
-			)
-			url = util.NodeURLJoin(is.api.Self().String(), "/RunState")
-			is.dchan <- core.NewEvent(
-				types.Event_DISCOVERY,
-				url,
-				&core.DiscoveryEvent{
-					URL:     url,
-					ValueID: cpb.Node_INIT.String(),
-				},
-			)
-		default:
-			is.api.Logf(types.LLERROR, "asked to perform unknown mutation: %s", me.Mutation[1])
-		}
-	case core.MutationEvent_INTERRUPT:
-		is.clearAllActions()
-		is.clearAllTriggers()
-		is.clearRunning()
-	}
-}
-
-// we only care about changes to containers themselves
-var reEvent = regexp.MustCompile(fmt.Sprintf(`^\/?%s\/Images/([A-Za-z.\-_]+)(\/?.*)$`, regexp.QuoteMeta(isURL)))
-
-// handle event handles the flood of all events and filters for state changes that we care about
-func (is *ImageAPI) handleEvent(m types.Event) {
-	if m.Type() != types.Event_STATE_CHANGE {
-		return
-	}
-	// we only care about self events having to do with imagestate
-	node, url := util.NodeURLSplit(m.URL())
-	if node != is.api.Self().String() {
-		// this event isn't about us
-		return
-	}
-	match := reEvent.FindAllStringSubmatch(url, -1)
-	if len(match) != 1 {
-		// not our url
-		return
-	}
-	name := match[0][1]
-	sub := match[0][2]
-	// ok, this is a self ImageState statechanmge event
-	sce := m.Data().(*core.StateChangeEvent)
-	switch sce.Type {
-	case cpb.StateChangeControl_CFG_UPDATE:
-		// container configuration changed
-		is.cfgChange(name, sub)
-	case cpb.StateChangeControl_UPDATE:
-		// container state changed
-		is.dscChange(name, sub, sce.Value)
-	}
-}
-
-// cfg changes general mark discover nodes as needing some kind of update
-func (is *ImageAPI) cfgChange(name, sub string) {
-	cfg, _ := is.api.QueryRead(is.api.Self().String())
-	dsc, _ := is.api.QueryReadDsc(is.api.Self().String())
-	ctnURL := util.URLPush(util.URLPush(isURL, "Images"), name)
-	if _, err := cfg.GetValue(ctnURL); err != nil {
-		// this Image is no longer configured.  It should be marked for deletion
-		if _, err := dsc.GetValue(ctnURL); err != nil {
-			// it seems to have already been deleted.  We're done.
-			is.updateSetState()
-			return
-		}
-		// mark node for update/delete
-		is.setValues(map[string]interface{}{
-			util.URLPush(ctnURL, "Action"): ia.Image_DELETE,
-			util.URLPush(ctnURL, "State"):  ia.ImageState_UPDATE,
-		})
-	} else {
-		_, err := dsc.GetValue(ctnURL)
-		if err != nil {
-			// this is a newly defined image
-			is.api.Logf(types.LLINFO, "new image definition found for %s", name)
-			is.setValues(map[string]interface{}{
-				ctnURL: ia.Image{
-					Container: &ia.Container{State: ContainerState_DELETED},
-					State:     ia.ImageState_UPDATE,
-					Action:    ia.Image_CREATE,
-				}})
-		} else {
-			// this Image needs to be updated/reloaded
-			is.setValues(map[string]interface{}{
-				util.URLPush(ctnURL, "Action"): ia.Image_RELOAD,
-				util.URLPush(ctnURL, "State"):  ia.ImageState_UPDATE,
-			})
-		}
-	}
-	is.updateSetState()
-}
-
-// updateSetState looks at all of the individual image states and derives the set state
-// it then discovers the state if it varies from the exiting one
-func (is *ImageAPI) updateSetState() {
-	v, _ := is.api.QueryGetValueDsc(is.api.Self().String(), isURL)
-	s := v.(ia.ImageSet)
-
-	state := float64(ia.ImageState_UNKNOWN)
-
-	for _, img := range s.Images {
-		// we rely on the numeric sequence here
-		state = math.Max(float64(img.State), state)
-	}
-
-	// this is a special case.  If there are no images, the state is whatever we want it to be.
-	v, _ = is.api.QueryGetValue(is.api.Self().String(), isURL)
-	cs := v.(ia.ImageSet)
-	if len(cs.Images) == 0 && len(s.Images) == 0 {
-		state = float64(is.target)
-	}
-
-	if ia.ImageState(state) != s.State {
-		// discover the imagestate update
-		// this always means we have to stop running
-		is.clearRunning()
-		is.clearAllActions()
-		is.clearAllTriggers()
-		is.dchan <- core.NewEvent(
-			types.Event_DISCOVERY,
-			util.NodeURLJoin(is.api.Self().String(), issURL),
-			&core.DiscoveryEvent{
-				URL:     util.NodeURLJoin(is.api.Self().String(), issURL),
-				ValueID: ia.ImageState(state).String(),
-			},
-		)
-	}
-}
-
-// dsc changes mean something about the container(s) changed
-func (is *ImageAPI) dscChange(name, sub string, value reflect.Value) {
-	// Note: currently sub always == "" because map value changes represent a single change
-	if sub == "/Action" {
-		is.activateImage(name)
-		return
-	}
-	if sub != "" && sub != "/State" && sub != "/Container/State" {
-		is.api.Logf(types.LLDDDEBUG, "unhandled dsc change path: %s", sub)
-	}
-	is.api.Logf(types.LLDDDEBUG, "processing dsc change path: %s", sub)
-	ctnURL := util.URLPush(util.URLPush(isURL, "Images"), name)
-	// ok, we had a change in running state
-	v, err := is.api.QueryGetValueDsc(is.api.Self().String(), ctnURL)
-	if err != nil || v == nil {
-		// image was deleted
-		is.fireTriggers(name, ContainerState_DELETED)
-		is.updateSetState()
-		return
-	}
-	idc := v.(ia.Image)
-
-	if idc.Container == nil || idc.Container.State == "" { // deleted
-		is.fireTriggers(name, ContainerState_DELETED)
-		is.updateSetState()
-		return
-	}
-
-	if is.fireTriggers(name, idc.Container.State) {
-		// we were expecting this change and we had triggers to fire
-		is.updateSetState()
-		return
-	}
-
-	if sub == "/Container/State" || sub == "" {
-		// there was a change in container state and we weren't expecting it (no triggers)
-		// Let's update the Image state accordingly
-		is.unexpectedStateChange(name)
-		is.updateSetState()
-		return
-	}
-
-	// Action may have been included in a whole-image change
-	if sub == "" && idc.Action != is.getAction(name) {
-		is.activateImage(name)
-		return
-	}
-}
-
-func (is *ImageAPI) unexpectedStateChange(name string) {
-	ctnURL := util.URLPush(util.URLPush(isURL, "Images"), name)
-	dsc, err := is.api.QueryGetValueDsc(is.api.Self().String(), ctnURL)
-	if err != nil {
-		return
-	}
-	idsc := dsc.(ia.Image)
-	cfg, err := is.api.QueryGetValue(is.api.Self().String(), ctnURL)
-	if err != nil {
-		return
-	}
-	icfg := cfg.(ia.Image)
-	target := is.target
-	is.api.Logf(types.LLERROR, "got an unexpected state change for image container %s: %s", name, idsc.Container.State)
-
-	if target == ia.ImageState_IDLE {
-		if idsc.Container.State != ContainerState_DELETED {
-			// we need to delete it
-			is.setAction(name, ia.Image_DELETE)
-			is.deleteImage(name, &icfg)
-			is.setTrigger(name, ContainerState_DELETED, is.clearAction)
-		}
-		return
-	}
-
-	switch idsc.Container.State {
-	case ContainerState_DEAD:
-		if idsc.State != ia.ImageState_ERROR && idsc.State != ia.ImageState_FATAL {
-			is.imageRaiseError(name, ia.Image_DIED)
-		}
-	case ContainerState_EXITED:
-		if icfg.Container.State != ContainerState_EXITED {
-			if idsc.State != ia.ImageState_ERROR && idsc.State != ia.ImageState_FATAL {
-				is.imageRaiseError(name, ia.Image_DIED)
-			}
-		}
-	case ContainerState_RUNNING:
-		// TODO: we should probably deal with unexpected running states
-	case ContainerState_STOPPING:
-	case ContainerState_CREATED:
-	}
-}
-
-// discover's job is to periodically poll the imageapi for current data and update discoverable state
-// discover does not make decisions on ImageState
-// ironically, discover uses SetValue + Update instead of discover, since it updates a complex data structure
-func (is *ImageAPI) discover() {
-	is.mutex.Lock()
-	is.api.Logf(types.LLDDDEBUG, "initiating image discovery")
-	iscv, _ := is.api.QueryGetValue(is.api.Self().String(), isURL)
-	isc := iscv.(ia.ImageSet)
-	isdv, _ := is.api.QueryGetValueDsc(is.api.Self().String(), isURL)
-	isd := isdv.(ia.ImageSet)
-	client := is.getAPIClient()
-	resp, err := client.Containers.ListContainers(containers.NewListContainersParams())
-	if err != nil {
-		// set error stuff
-		is.mutex.Unlock()
-		if cerr, ok := err.(*containers.ListContainersDefault); ok {
-			is.api.Logf(types.LLERROR, "failed to list containers: Code: %d Message: %s", cerr.Payload.Code, *cerr.Payload.Message)
-		} else {
-			is.api.Logf(types.LLERROR, "failed to call imageapi: %v", err)
-		}
-		return
-	}
-	ctnsList := resp.GetPayload()
-
-	// dset is where we will build the ImageSet that we will discover for the node
-	is.api.Logf(types.LLDDDEBUG, "discover found %d containers", len(ctnsList))
-	dset := &ia.ImageSet{
-		State:  isd.State,              // we preserve state values
-		Images: map[string]*ia.Image{}, // but start with an empty list
-	}
-	for _, ca := range ctnsList {
-		cp := is.apiToProto(ca)
-		if cp == nil { // if conversion fails, don't crash, just report an error
-			is.api.Logf(types.LLERROR, "apiToProto conversion failed for container object")
-			continue
-		}
-		if ds, ok := isd.Images[cp.Name]; ok {
-			// this is one we already knew
-			ds.Container = cp
-			dset.Images[cp.Name] = ds
-			delete(isc.Images, cp.Name) // delete it so we know what we *didn't* find later
-		} else {
-			// this is a new one
-			dset.Images[cp.Name] = &ia.Image{
-				Container: cp,
-			}
-		}
-	}
-	// we now have a list of everything imageapi knows about...
-	// we also want at least a stub for anything in the cfg set
-	for name, cp := range isc.Images {
-		action := ia.Image_CREATE
-		if cur, ok := is.actions[name]; ok && cur == ia.Image_RELOAD {
-			// don't clobber the reload action
-			action = ia.Image_RELOAD
-		}
-		if cp.Container.State != ContainerState_RUNNING {
-			action = ia.Image_NONE
-		}
-		if _, ok := dset.Images[name]; !ok {
-			// we don't have an entry, does dsc?
-			if ds, ok := isd.Images[name]; ok {
-				ds.Container = &ia.Container{State: ContainerState_DELETED}
-				ds.Action = action
-				dset.Images[name] = ds
-			} else {
-				// nope, make a stub
-				dset.Images[name] = &ia.Image{
-					Container: &ia.Container{State: ContainerState_DELETED},
-					State:     ia.ImageState_IDLE,
-					Action:    action,
-				}
-			}
-		}
-	}
-
-	// ok, we're ready to ship it
-	// we do a query instead of a discover because we're setting various values
-	is.api.QuerySetValueDsc(is.api.Self().String(), isURL, *dset)
-	is.mutex.Unlock()
-}
-
-////////////////////////
-// Utility functions //
-//////////////////////
-
-// get a configured API client
-func (is ImageAPI) getAPIClient() *api.Imageapi {
-	as := is.cfg.ApiServer
-	t := transport.New(fmt.Sprintf("%s:%d", as.Server, as.Port), as.ApiBase, []string{"http"})
-	return api.New(t, strfmt.Default)
-}
-
-// not super efficient, but should work
-func (is *ImageAPI) protoToAPI(p *ia.Container) *models.Container {
-	b, err := json.Marshal(p)
-	if err != nil {
-		is.api.Logf(types.LLFATAL, "error translating container proto -> API: %v", err)
-		return nil
-	}
-	r := &models.Container{}
-	err = json.Unmarshal(b, r)
-	if err != nil {
-		is.api.Logf(types.LLFATAL, "error translating container proto -> API: %v", err)
-		return nil
-	}
-	return r
-}
-
-// not super efficient, but should work
-func (is *ImageAPI) apiToProto(a *models.Container) *ia.Container {
-	b, err := json.Marshal(a)
-	if err != nil {
-		is.api.Logf(types.LLFATAL, "error translating container API -> proto: %v", err)
-		return nil
-	}
-	r := &ia.Container{}
-	kjson.Unmarshal(b, r)
-	if err != nil {
-		is.api.Logf(types.LLFATAL, "error translating container API -> proto: %v", err)
-		return nil
-	}
-	return r
-}
-
-func (is *ImageAPI) imageRaiseError(name string, e ia.Image_ErrorCode) {
-	url := util.URLPush(util.URLPush(isURL, "Images"), name)
-	is.setValues(map[string]interface{}{
-		util.URLPush(url, "State"):     ia.ImageState_ERROR,
-		util.URLPush(url, "LastError"): e,
-	})
-}
-
-func (is *ImageAPI) createImage(name string, image *ia.Image) {
-	is.api.Logf(types.LLINFO, "creating image %s", name)
-	// make sure naming is uniform.  The container must be named the same as the map.
-	image.Container.Name = name
-	client := is.getAPIClient()
-	apiContainer := is.protoToAPI(image.Container)
-	params := containers.NewCreateContainerParams()
-	params.Container = apiContainer
-	_, err := client.Containers.CreateContainer(params)
-	if err != nil {
-		if cerr, ok := err.(*containers.CreateContainerDefault); ok {
-			is.api.Logf(types.LLERROR, "container creation failed for image %s: Code: %d Message: %s", name, cerr.Payload.Code, *cerr.Payload.Message)
-			// "discover" our error state
-		} else {
-			is.api.Logf(types.LLERROR, "failed to call imageapi: %v", err)
-		}
-		is.imageRaiseError(name, ia.Image_ATTACH)
-	}
-}
-
-func (is *ImageAPI) deleteImage(name string, image *ia.Image) {
-	url := isURL
-	for _, u := range []string{"Images", name, "Container", "State"} {
-		url = util.URLPush(url, u)
-	}
-	v, err := is.api.QueryGetValueDsc(is.api.Self().String(), url)
-	if err != nil || v.(string) == ContainerState_DELETED {
-		// image must not exist, so our work is done
-		is.api.Logf(types.LLDDDEBUG, "refusing to delete an already deleted image: %s", name)
-		return
-	}
-	client := is.getAPIClient()
-	switch v.(string) {
-	case ContainerState_RUNNING:
-		// tell the node to stop
-		is.api.Logf(types.LLINFO, "stopping image %s", name)
-		params := containers.NewSetContainerStateBynameParams()
-		params.Name = name
-		params.State = string(models.ContainerStateExited)
-		_, err := client.Containers.SetContainerStateByname(params)
-		if err != nil {
-			if cerr, ok := err.(*containers.SetContainerStateBynameDefault); ok {
-				is.api.Logf(types.LLERROR, "failed to stop container %s: Code: %d Message %s", name, cerr.Payload.Code, *cerr.Payload.Message)
-			} else {
-				is.api.Logf(types.LLERROR, "failed to call imageapi: %v", err)
-			}
-			// "discover" our error state
-			is.imageRaiseError(name, ia.Image_ATTACH)
-		}
-		fallthrough
-	case ContainerState_STOPPING:
-		// set a trigger and get rerun when it's done
-		is.api.Logf(types.LLINFO, "waiting for image %s to exit", name)
-		is.setTrigger(name, ContainerState_EXITED, is.tDelete)
-		return
-	}
-	// ok, we're ready to delete
-	is.api.Logf(types.LLINFO, "deleting image %s", name)
-	params := containers.NewDeleteContainerBynameParams()
-	params.Name = name
-	_, err = client.Containers.DeleteContainerByname(params)
-	if err != nil {
-		if cerr, ok := err.(*containers.DeleteContainerBynameDefault); ok {
-			is.api.Logf(types.LLERROR, "failed to delete container %s: Code: %d Message %s", name, cerr.Payload.Code, *cerr.Payload.Message)
-			// "discover" our error state
-		} else {
-			is.api.Logf(types.LLERROR, "failed to call imageapi: %v", err)
-		}
-		is.imageRaiseError(name, ia.Image_ATTACH)
-	}
-}
-
-func (is *ImageAPI) reloadImage(name string, image *ia.Image) {
-	is.api.Logf(types.LLINFO, "reloading image %s", name)
-	is.setTrigger(name, ContainerState_DELETED, is.tCreate)
-	is.deleteImage(name, image)
-}
-
-///////////////////////
-// trigger handling //
-/////////////////////
-
-// Triggers are hooks that get called when a certain ContainerState is reached
-// This allows for chaining of async actions
-// All triggers take an Image name as argument
-
-// we define an extra container state to recognize deleted containers
-const (
-	ContainerState_DELETED  = "DELETED"
-	ContainerState_CREATED  = "created"
-	ContainerState_RUNNING  = "running"
-	ContainerState_STOPPING = "stopping"
-	ContainerState_EXITED   = "exited"
-	ContainerState_DEAD     = "dead"
-)
-
-func (is *ImageAPI) clearAllTriggers() {
-	is.api.Logf(types.LLDDDEBUG, "clearing all triggers")
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	is.triggers = make(map[string]map[string][]trigger)
-}
-
-func (is *ImageAPI) clearTriggers(name string) {
-	is.api.Logf(types.LLDDDEBUG, "clearing triggers for %s", name)
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	is.triggers[name] = make(map[string][]trigger)
-}
-
-func (is *ImageAPI) setTrigger(name, state string, t trigger) {
-	is.api.Logf(types.LLDDDEBUG, "setting trigger %s for %s", state, name)
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	var ok bool
-	var ts map[string][]trigger
-	if ts, ok = is.triggers[name]; !ok {
-		is.triggers[name] = make(map[string][]trigger)
-		ts = is.triggers[name]
-	}
-	ts[state] = append(ts[state], t)
-}
-
-func (is *ImageAPI) fireTriggers(name, state string) bool {
-	is.mutex.Lock()
-	if ts, ok := is.triggers[name][state]; ok {
-		if len(ts) == 0 {
-			is.mutex.Unlock()
-			return false
-		}
-		is.api.Logf(types.LLDDEBUG, "firing image %d triggers %s for %s", len(ts), state, name)
-		delete(is.triggers[name], state)
-		is.mutex.Unlock()
-		for _, t := range ts {
-			t(name)
-		}
-		return true
-	}
-	is.mutex.Unlock()
-	return false
-}
-
-func (is *ImageAPI) tDelete(name string) {
-	// we use an empty image
-	time.Sleep(5 * time.Second) // FIXME: this is a hack to give rbd/fs time to flush.  This should be fixed in imageapi-server
-	is.deleteImage(name, &ia.Image{})
-}
-
-func (is *ImageAPI) tCreate(name string) {
-	iurl := util.URLPush(util.URLPush(isURL, "Images"), name)
-	v, err := is.api.QueryGetValue(is.api.Self().String(), iurl)
-	image := v.(ia.Image)
-	if err != nil {
-		// apparently there's nothing to create
-		is.api.Logf(types.LLDEBUG, "triggered to create image that is no longer defined: %s", name)
-		return
-	}
-	is.createImage(name, &image)
-}
-
-func (is *ImageAPI) tSetACTIVE(name string) {
-	is.api.Logf(types.LLINFO, "image %s becomes ACTIVE", name)
-	iurl := util.URLPush(util.URLPush(isURL, "Images"), name)
-	is.setValues(map[string]interface{}{
-		util.URLPush(iurl, "State"):   ia.ImageState_ACTIVE,
-		util.URLPush(iurl, "Action"):  ia.Image_NONE,
-		util.URLPush(iurl, "Retries"): int32(0),
-	})
-}
-
-func (is *ImageAPI) tSetIDLE(name string) {
-	is.api.Logf(types.LLINFO, "image %s becomes IDLE", name)
-	iurl := util.URLPush(util.URLPush(isURL, "Images"), name)
-	is.setValues(map[string]interface{}{
-		util.URLPush(iurl, "State"):   ia.ImageState_IDLE,
-		util.URLPush(iurl, "Action"):  ia.Image_CREATE,
-		util.URLPush(iurl, "Retries"): int32(0),
-	})
-}
-
-// action handling
-
-func (is *ImageAPI) setRunning() {
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	is.running = true
-}
-
-func (is *ImageAPI) isRunning() bool {
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	return is.running
-}
-
-func (is *ImageAPI) clearRunning() {
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	is.running = false
-}
-
-func (is *ImageAPI) setAction(name string, action ia.Image_Action) {
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	is.actions[name] = action
-}
-
-// note: also a trigger
-func (is *ImageAPI) clearAction(name string) {
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	if _, ok := is.actions[name]; ok {
-		delete(is.actions, name)
-	}
-}
-
-func (is *ImageAPI) clearAllActions() {
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	is.actions = make(map[string]ia.Image_Action)
-}
-
-func (is *ImageAPI) getAction(name string) ia.Image_Action {
-	is.mutex.Lock()
-	defer is.mutex.Unlock()
-	if a, ok := is.actions[name]; ok {
-		return a
-	}
-	return ia.Image_NONE
 }
